% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_tips.r
\name{common_tips}
\alias{common_tips}
\title{Find shared tip labels}
\usage{
common_tips(x)
}
\arguments{
\item{x}{a collection of trees of class "multiPhylo"}
}
\value{
intersection of tip labels across all trees
}
\description{
Find shared tip labels of a collection of trees.
This function is useful when pruning a collection
of related trees to a trees of same size and taxa.
}
\examples{
trees = rankedPhylo(3:7)
common_tips(trees)

}
