% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_newick.r
\name{read_newick}
\alias{read_newick}
\title{Read a newick tree}
\usage{
read_newick(x)
}
\arguments{
\item{x}{a single character string containing tree in a newick format}
}
\value{
object of class "rankedPhylo"
}
\description{
Read a tree in a newick format and convert it to a rankedPhylo
}
\examples{
read_newick("((A:1,B:1):1,C:2);")

# Note, not all valid newick trees are valid ranked trees
try( read_newick("((A,B),C);") ) 

}
