% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnni.r
\name{rnni}
\alias{rnni}
\title{Calculate RNNNI distance}
\usage{
rnni(x, y)
}
\arguments{
\item{x}{a tree of class "rankedPhylo"}

\item{y}{a tree of class "rankedPhylo"}
}
\value{
an integer RNNI distance between \code{x} and \code{y}
}
\description{
Calculate RNNNI distance
}
