% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrnni.r
\docType{package}
\name{rrnni}
\alias{rrnni}
\title{rrnni: Calculate a distance and a path in RNNI space}
\description{
TODO longer description
}
\references{
\cite{
Gavryushkin, A., Whidden, C., Matsen, IV. F. A. (2018).
The combinatorics of discrete time-trees: theory and open problems.
Journal of Mathematical Biology, 76(5):1101-1121.
doi:10.1007/s00285-017-1167-9, \url{https://doi.org/10.1007/s00285-017-1167-9}
}

\cite{
Collienne, L. and Gavryushkin A. (2021).
Computing Nearest Neighbour Interchange Distances Between Ranked Phylogenetic Trees.
Journal of Mathematical Biology, 82(1-2):8.
doi:10.1007/s00285-021-01567-5, \url{https://doi.org/10.1007/s00285-021-01567-5}
}

\cite{ Collienne, L. (2021). Spaces of Phylogenetic Time Trees. University of Otago. }
}
