% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rankedPhylo.r
\name{validate_rankedPhylo}
\alias{validate_rankedPhylo}
\title{Validate rankedPhylo class}
\usage{
validate_rankedPhylo(x)
}
\arguments{
\item{x}{an object of class "rankedPhylo"}
}
\value{
This function doesn't have a return value.
}
\description{
Validate that the object is a rankedPhylo object with a proper structure and type.
}
\details{
This is internal function used for testing purposes.

"rankedPhylo" is a subclass of "phylo" object with several restrictions given by the structure
of the represented ranked tree.
\code{validate_rankedPhylo} checks these restrictions and throw a particular error, if the object
fails any of these restrictions.
}
\keyword{internal}
