% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmv.r
\name{rmv}
\alias{rmv}
\title{Remove values from a vector}
\usage{
rmv(x, values)
}
\arguments{
\item{x}{a vector}

\item{values}{vector of values}
}
\value{
x without values
}
\description{
Remove one or more values from a vector.
This is a shorthand for \code{x[!x \%in\% values]}
}
\keyword{internal}
